fix: table group render
